client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["gavril-s <gavril.senkevich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
common = {path = "../common", develop=true}


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
